+++ Client.Test/WriteApiTest.cs
@@ -20,7 +20,7 @@ namespace InfluxDB.Client.Test
         [SetUp]
         public new void SetUp()
         {
-            _influxDbClient = InfluxDBClientFactory.Create(MockServerUrl, "token".ToCharArray());
+            _influxDbClient = InfluxDBClientFactory.Create(MockServerUrl, "token");
             _writeApi = _influxDbClient.GetWriteApi();
         }